Fix the Dataproc handlers and pytests
