+++ dataproc_jupyter_plugin/controllers/dataproc.py
@@ -14,6 +14,7 @@
 
 import json
 
+import aiohttp
 import tornado
 from jupyter_server.base.handlers import APIHandler
 
@@ -27,8 +28,11 @@ class ClusterListController(APIHandler):
         try:
             page_token = self.get_argument("pageToken")
             page_size = self.get_argument("pageSize")
-            client = dataproc.Client(await credentials.get_cached(), self.log)
-            cluster_list = await client.list_clusters(page_size, page_token)
+            async with aiohttp.ClientSession() as client_session:
+                client = dataproc.Client(
+                    await credentials.get_cached(), self.log, client_session
+                )
+                cluster_list = await client.list_clusters(page_size, page_token)
             self.finish(json.dumps(cluster_list))
         except Exception as e:
             self.log.exception("Error fetching cluster list")
@@ -41,8 +45,11 @@ class RuntimeController(APIHandler):
         try:
             page_token = self.get_argument("pageToken")
             page_size = self.get_argument("pageSize")
-            client = dataproc.Client(await credentials.get_cached(), self.log)
-            runtime_list = await client.list_runtime(page_size, page_token)
+            async with aiohttp.ClientSession() as client_session:
+                client = dataproc.Client(
+                    await credentials.get_cached(), self.log, client_session
+                )
+                runtime_list = await client.list_runtime(page_size, page_token)
             self.finish(json.dumps(runtime_list))
         except Exception as e:
             self.log.exception(f"Error fetching runtime template list: {str(e)}")
