+++ dataproc_jupyter_plugin/tests/test_bigquery.py
@@ -13,73 +13,15 @@
 # limitations under the License.
 
 import json
-from unittest.mock import Mock
 
-import aiohttp
 from google.cloud import jupyter_config
 
 from dataproc_jupyter_plugin import credentials
-
-
-async def mock_credentials():
-    return {
-        "project_id": "credentials-project",
-        "project_number": 12345,
-        "region_id": "mock-region",
-        "access_token": "mock-token",
-        "config_error": 0,
-        "login_error": 0,
-    }
-
-
-class MockResponse():
-    def __init__(self, json):
-        self._json = json
-        self.status = 200
-    
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, *args, **kwargs):
-        return
-
-    async def json(self):
-        return self._json
-
-
-class MockClientSession():
-    async def __aenter__(self):
-        return self
-
-    async def __aexit__(self, *args, **kwargs):
-        return
-
-    def get(self, api_endpoint, headers=None):
-        return MockResponse({
-            "api_endpoint": api_endpoint,
-            "headers": headers,
-        })
-
-    def post(self, api_endpoint, headers=None, json=None):
-        return MockResponse({
-            "results": [
-                {
-                    "api_endpoint": api_endpoint,
-                    "headers": headers,
-                    "json": json,
-                }
-            ]
-        })
-
-
-async def mock_config(field_name):
-    return None
+from dataproc_jupyter_plugin.tests import mocks
 
 
 async def test_list_datasets(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp,  "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     mock_project_id = "mock-project-id"
     mock_page_token = "mock-page-token"
@@ -101,9 +43,8 @@ async def test_list_datasets_with_invalid_credentials(monkeypatch, jp_fetch):
     async def mock_credentials():
         return {}
 
+    mocks.patch_mocks(monkeypatch)
     monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
     response = await jp_fetch(
         "dataproc-plugin",
         "bigQueryDataset",
@@ -115,9 +56,7 @@ async def test_list_datasets_with_invalid_credentials(monkeypatch, jp_fetch):
 
 
 async def test_list_tables(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     mock_dataset_id = "mock-dataset-id"
     mock_project_id = "mock-project-id"
@@ -141,9 +80,7 @@ async def test_list_tables(monkeypatch, jp_fetch):
 
 
 async def test_dataset_info(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     mock_dataset_id = "mock-dataset-id"
     mock_project_id = "mock-project-id"
@@ -165,9 +102,7 @@ async def test_dataset_info(monkeypatch, jp_fetch):
 
 
 async def test_table_info(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     mock_dataset_id = "mock-dataset-id"
     mock_project_id = "mock-project-id"
@@ -191,9 +126,7 @@ async def test_table_info(monkeypatch, jp_fetch):
 
 
 async def test_preview(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     mock_dataset_id = "mock-dataset-id"
     mock_max_results = "mock-max-results"
@@ -221,8 +154,7 @@ async def test_preview(monkeypatch, jp_fetch):
 
 
 async def test_projects_list(monkeypatch, jp_fetch):
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
+    mocks.patch_mocks(monkeypatch)
 
     response = await jp_fetch(
         "dataproc-plugin",
@@ -237,10 +169,8 @@ async def test_search(monkeypatch, jp_fetch):
     async def mock_config(config_field):
         return ""
 
+    mocks.patch_mocks(monkeypatch)
     monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(credentials, "get_cached", mock_credentials)
-    monkeypatch.setattr(jupyter_config, "async_get_gcloud_config", mock_config)
-    monkeypatch.setattr(aiohttp, "ClientSession", MockClientSession)
 
     mock_search_string = "mock-search-string"
     mock_system = "mock-system"
